backend/pyproject.toml
[tool.poetry]
name = "jetfork-expertise-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.97.0"
uvicorn = "^0.22.0"
pydantic = "^1.10.9"
requests = "^2.31.0"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.6"
asyncpg = "^0.27.0"
SQLAlchemy = "^2.0.16"
gunicorn = "^20.1.0"
fastapi-users = { extras = ["sqlalchemy"], version = "^11.0.0" }
minio = "^7.1.15"
python-dotenv = "^1.0.0"
celery = { extras = ["redis"], version = "^5.3.1" }
"pdfminer.six" = "20221105"
wrapt_timeout_decorator = "1.3.12.2"
rapidfuzz = "3.1.1"
textdistance = "4.5.0"
catboost = "1.2"
joblib = "^1.2.0"
pypdf2 = "3.0.1"
nltk = "^3.8.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.2"
ipython = "^8.14.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.cruft]
skip = [".env"]

[tool.pytest.ini_options]
asyncio_mode = 'auto'
